requirements.txt
cssselect2==0.5.0
ffmpeg-python==0.2.0
future==0.18.2
get-video-properties==0.1.1
lxml==4.8.0
Pillow==9.0.1
reportlab==3.6.8
svglib==1.2.1
tinycss2==1.1.1
webencodings==0.5.1
